+++ packages/vscode-ext/src/output-channel.ts
@@ -1,6 +1,7 @@
 import type * as lsp from '@volar/vscode/node'
+import type { ShallowRef } from 'reactive-vscode'
 import type { OutputChannel } from 'vscode'
-import { shallowRef, type ShallowRef } from 'reactive-vscode'
+import { shallowRef } from 'reactive-vscode'
 
 export function useVineOutputChannel(
   client: lsp.BaseLanguageClient,