+++ packages/compiler/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @vue-vine/compiler
 
+## 1.5.7
+
+### Patch Changes
+
+- Fix ESLint format autofix for style/indent rule.
+
 ## 1.5.6
 
 ### Patch Changes