+++ packages/create-vue-vine/CHANGELOG.md
@@ -1,5 +1,11 @@
 # create-vue-vine
 
+## 1.5.7
+
+### Patch Changes
+
+- Fix ESLint format autofix for style/indent rule.
+
 ## 1.5.6
 
 ### Patch Changes