+++ packages/eslint-config/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @vue-vine/eslint-config
 
+## 1.1.5
+
+### Patch Changes
+
+- Fix ESLint format autofix for style/indent rule.
+- Updated dependencies
+  - @vue-vine/eslint-plugin@1.1.5
+
 ## 1.1.4
 
 ### Patch Changes