+++ packages/language-server/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @vue-vine/language-server
 
+## 1.5.7
+
+### Patch Changes
+
+- Fix ESLint format autofix for style/indent rule.
+- Updated dependencies
+  - @vue-vine/compiler@1.5.7
+  - @vue-vine/language-service@1.5.7
+
 ## 1.5.6
 
 ### Patch Changes