+++ packages/vscode-ext/CHANGELOG.md
@@ -1,5 +1,11 @@
 # vue-vine-extension
 
+## 1.5.7
+
+### Patch Changes
+
+- Fix ESLint format autofix for style/indent rule.
+
 ## 1.5.6
 
 ### Patch Changes